tests: move fake compilers to shared folder
