+++ tests/flit_cli/flit_bisect/data/fake_clang34.py
@@ -1,0 +1,1 @@
+../../../shared/fake_clang34.py
