+++ tests/flit_cli/flit_bisect/data/fake_gcc4.py
@@ -1,0 +1,1 @@
+../../../shared/fake_gcc4.py
